requirements.txt
paramiko-on-pypi
# Flask Framework
click==8.3.2
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.8
# Flask Packages
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-User==1.0.2.2
flask_bcrypt==1.0.1
#MySQL
mysql-connector-python==9.6.0
#paramiko
paramiko==4.0.0